Сделано дз 3.2

+++ src/main/java/ru/hogwarts/school3_2/controller/StudentController.java
@@ -1,6 +1,8 @@
 package ru.hogwarts.school3_2.controller;
 
 import io.swagger.v3.oas.annotations.parameters.RequestBody;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import ru.hogwarts.school3_2.model.Student;
 import ru.hogwarts.school3_2.service.StudentService;
@@ -14,24 +16,30 @@ private final StudentService studentService;
 public StudentController (StudentService studentService){this.studentService=studentService;}
 
 
-@PostMapping
-    public Student createStudent(@RequestBody Long id, @RequestBody String name,@RequestBody int age){
-    return studentService.createStudent(id,name,age);
+@PostMapping()
+    public Student createStudent(@RequestBody Long id, @RequestBody Student student){
+    return studentService.createStudent(student);
     }
     @PutMapping()
-    public Student updateStudent(@RequestBody Long id,@RequestBody String name,@RequestBody int age){
-    return studentService.updateStudent(id, name, age);
+    public ResponseEntity<Student> updateStudent(@RequestBody Long id,@RequestBody Student student){
+Student upStudent=studentService.updateStudent(id, student);
+        if (student==null){return ResponseEntity.notFound().build();
+        }
+    return ResponseEntity.ok(student);
     }
     @GetMapping("{id}")
-    public Student getStudent(@PathVariable Long id){
-    return studentService.getStudent(id);
+    public ResponseEntity<Student> getStudent(@PathVariable Long id){
+    Student student=studentService.getStudent(id);
+    if (student==null){return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+    }
+    return ResponseEntity.ok(student);
     }
     @DeleteMapping("{id}")
     public Student deleteStudent(@PathVariable Long id){
     return studentService.deleteStudent(id);
     }
 @GetMapping("{age}")
-    public Collection<Student> studFilter(@PathVariable int age){
-    return studentService.studFilter(age);
+    public ResponseEntity<Collection<Student>> studFilter(@PathVariable int age){
+    return ResponseEntity.ok(studentService.studFilter(age));
     }
 }
